+++ RELEASE NOTES.md
@@ -3,6 +3,7 @@ Version 1.5 beta
 - Added innerShadowBlur property
 - Added characterSpacing property
 - Added kerningTable property for per-character spacing
+- Added baselineOffset property
 - LineSpacing property is now relative to pointSize
 
 Version 1.4.2